requirements.txt
streamlit>=1.30.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.3.0
google-genai>=1.0.0
pymupdf4llm>=0.0.10
pymupdf>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
markdownify>=0.12.0
rank_bm25>=0.2.2
numpy>=1.24.0
